% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post}
\alias{post}
\title{A Hugo markdown format}
\usage{
post(fig_width = 7, fig_height = 5, fig_retina = NULL, dev = "png",
  includes = NULL, md_extensions = NULL, pandoc_args = NULL)
}
\description{
An output format suitable for subsequent processing by hugo's Blackfriday.
}
\details{
Unlike the default markdown format, YAML is always preserved and no TOC is
ever generated.
}

